backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==10.2.0
ultralytics>=8.2.0
opencv-python-headless==4.9.0.80
numpy<2.0.0
torch>=2.2.0
